implement suggestions received in review
